requirements.txt
google-cloud-bigquery==3.11.4
pandas==2.1.1
pyarrow==13.0.0